+++ core/network/src/main/java/com/school_of_company/network/datasource/expo/ExpoDataSource.kt
@@ -1,14 +1,14 @@
 package com.school_of_company.network.datasource.expo
 
-import com.school_of_company.network.dto.expo.all.ExpoRequestAndResponse
+import com.school_of_company.network.dto.expo.response.ExpoResponse
 import com.school_of_company.network.dto.expo.response.ExpoIdResponse
 import com.school_of_company.network.dto.expo.response.ExpoListResponse
 import kotlinx.coroutines.flow.Flow
 
 interface ExpoDataSource {
     fun getExpoList() : Flow<List<ExpoListResponse>>
-    fun getExpoInformation(expoId: String) : Flow<ExpoRequestAndResponse>
-    fun registerExpoInformation(body: ExpoRequestAndResponse) : Flow<ExpoIdResponse>
-    fun modifyExpoInformation(expoId: String, body: ExpoRequestAndResponse) : Flow<Unit>
+    fun getExpoInformation(expoId: String) : Flow<ExpoResponse>
+    fun registerExpoInformation(body: ExpoResponse) : Flow<ExpoIdResponse>
+    fun modifyExpoInformation(expoId: String, body: ExpoResponse) : Flow<Unit>
     fun deleteExpoInformation(expoId: String) : Flow<Unit>
 }
