+++ core/network/src/main/java/com/school_of_company/network/dto/expo/request/ExpoAllRequest.kt
@@ -1,0 +1,3 @@
+package com.school_of_company.network.dto.expo.request
+
+data class ExpoAllRequest()
