+++ feature/signin/src/main/java/com/school_of_company/viewmodel/SignInViewModel.kt
@@ -1,6 +1,5 @@
 package com.school_of_company.signin.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
